Added no-validations, no-repository options to add service command
