+++ lib/addToLocalNuget.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-# dotnet nuget remove source "tada_project"
-# dotnet nuget add source --name "tada_project" "C:\Users\n_dua\source\tada\output"
+dotnet nuget remove source "tada_project"
+dotnet nuget add source --name "tada_project" "C:\Users\n_dua\source\tada\output"
