+++ lib/installLocally.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-version=1.0.0
+version=1.0.0-alpha
 
 dotnet pack;
 dotnet tool uninstall Tada --global;
