--- src/Tada.TemplatePack/templates/service/full/.template.config/template.json
@@ -1,29 +1,0 @@
-{
-  "$schema": "http://json.schemastore.org/template",
-  "author": "Niall",
-  "classifications": [
-    "Tada service domain",
-    "clean architecture"
-  ],
-  "identity": "Tada.service.full",
-  "name": "Tada.service.full",
-  "shortName": "tada-service-full",
-  "sourceName": "TadaTemplateName",
-  "tags": {
-    "language": "C#",
-    "type": "item"
-  },
-  "description": "Creates a Service Items in an Api Solution",
-  "preferNameDirectory": false,
-  "symbols": {
-    "nameSpace": {
-      "description": "service namespace",
-      "dataType": "string",
-      "type": "parameter",
-      "defaultValue": "tada",
-      "isRequired": true,
-      "replaces": "TadaSourceName",
-      "fileRename": "TadaSourceName"
-    }
-  }
-}
