--- src/Tada.TemplatePack/templates/service/full/src/3.Services/TadaSourceName.Services.Tests/TadaTemplateNames/DeleteTadaTemplateNameServiceTests.cs
@@ -1,23 +1,0 @@
-using Moq;
-
-namespace TadaSourceName.Services.Tests.TadaTemplateNames;
-
-public class DeleteTadaTemplateNameServiceTests : TadaTemplateNameServiceFixture
-{
-
-    public DeleteTadaTemplateNameServiceTests() : base()
-    {
-
-    }
-
-    [Fact]
-    public async Task DeleteTadaTemplateName_Success()
-    {
-        var expected = Guid.NewGuid();
-        mockTadaTemplateNameRepository.Setup(x => x.Delete(It.IsAny<Guid>()));
-        var result = await context.DeleteTadaTemplateName(expected.ToString());
-
-        Assert.NotNull(result);
-        //todo: check that delete was successful
-    }
-}
